analysis working with git ignore
